refactor: 단체장 API, 단체 공통 API 분리 #59
